% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentdirections.R
\name{plot_svd}
\alias{plot_svd}
\title{Plot data using the first and second singular vector}
\usage{
plot_svd(a, label = NULL, b = a)
}
\arguments{
\item{a}{Original matrix to be sketched (n x m)}

\item{label}{Group index for each a's row. These values are used for group and color.}

\item{b}{A sketched matrix (l x m)}
}
\description{
Plot data using the first and second singular vector
}
