% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentdirections.R
\name{sketching}
\alias{sketching}
\title{Compute a sketch matrix of input matrix}
\usage{
sketching(a, l, eps = 10^(-8))
}
\arguments{
\item{a}{Original matrix to be sketched (n x m)}

\item{l}{The number of rows in sketched matrix (l x m)}

\item{eps}{If a value is smaller than eps, that is considered as equal to zero. The default value is 10^(-8)}
}
\description{
Compute a sketch matrix of input matrix
}
